+++ src/@components/@common/Input.tsx
@@ -75,11 +75,15 @@ const InputWrapper = styled.input<{ width: number; color: string | undefined }>`
 
   color: white;
 
-  border-bottom: 1px solid ${({ color }) => color};
+  border-bottom: 1px solid ${({ theme, color }) => (color === undefined ? theme.colors.gray3 : color)};
 
   width: ${({ width }) => width}rem;
 
   ${({ theme }) => theme.fonts.input}
+
+  &:focus {
+    border-color: ${({ theme, color }) => (color === undefined ? theme.colors.white : color)};
+  }
 `;
 
 const ErrorMessage = styled.h1<{ color: string }>`
