+++ src/@components/@common/editLayout.tsx
@@ -1,0 +1,32 @@
+import { PropsWithChildren } from "react";
+import styled from "styled-components";
+
+export default function EditLayout(props: PropsWithChildren) {
+  const { children } = props;
+
+  return <EditBox>{children}</EditBox>;
+}
+
+const EditBox = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  position: absolute;
+  right: 18.1rem;
+
+  width: 77.9rem;
+  height: 88.8rem;
+
+  margin-left: 10rem;
+
+  backdrop-filter: blur(1rem);
+
+  border: 0.3rem solid transparent;
+  border-radius: 5rem;
+  background-image: linear-gradient(rgba(13, 14, 17, 0.9), rgba(20, 21, 23, 0.6)),
+    linear-gradient(to top, transparent, #3e4045);
+
+  background-origin: border-box;
+  background-clip: content-box, border-box;
+`;
