+++ src/@components/signUp/index.tsx
@@ -6,6 +6,7 @@ import { SIGNUP_STEP } from "../../core/signUp/stepRenderer";
 import useConventionModal from "../../hooks/common/useConventionModal";
 import { role } from "../../recoil/common/role";
 import ConventionModal from "../@common/conventionModal";
+import EditLayout from "../@common/editLayout";
 import Footer from "../@common/footer";
 import SignUpBackButton from "./signUpBackButton";
 import StepFooter from "./stepFooter";
@@ -30,11 +31,11 @@ export default function SignupStep() {
       <SignUpContainer>
         <Img src={background} alt="배경" />
 
-        <StepBox>
+        <EditLayout>
           <StepHeader step={step} />
           <StepMain step={step} />
           <StepFooter step={step} setStep={setStep} />
-        </StepBox>
+        </EditLayout>
       </SignUpContainer>
       <Footer />
     </>
@@ -54,28 +55,4 @@ const Img = styled.img`
   position: absolute;
   width: 192rem;
   height: 98rem;
-`;
-
-const StepBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  position: absolute;
-  right: 18.1rem;
-
-  width: 77.9rem;
-  height: 88.8rem;
-
-  margin-left: 10rem;
-
-  backdrop-filter: blur(1rem);
-
-  border: 0.3rem solid transparent;
-  border-radius: 5rem;
-  background-image: linear-gradient(rgba(13, 14, 17, 0.9), rgba(20, 21, 23, 0.6)),
-    linear-gradient(to top, transparent, #3e4045);
-
-  background-origin: border-box;
-  background-clip: content-box, border-box;
 `;
